if i use this pattern, how do i fetch the previous value
I think the obvious thing is to introduce more state within the parser.
I think that makes it look more like the "eatToken" approach. I'm still
feeling like the array of nodes at each level is easier to work with
than having a more complex pointer structure, but I need to get some
helper methods in here somewhere to make the pointer-like operation
easier. Perhaps introducing some kind of program-counter vector that is
3 dimensional. Actually, is it 3 dimensional or just 2?

Still debating if I shouldn't try to better understand how to do a more
malloc-y style thing, but I think the efficiency of the arrays and using
the string representations with regexes and looking a bit more like a
formal grammar is something to keep trying.